refactor: make 'ListArticles' component reusable
